Добавляет контроллеры
